observations: move 6 forms to Views/
Per #4331. Six single-use observation forms move; two flagged in the
audit (LightboxObservationTitle, MatrixBoxTitle) stay because they
also render from reusable components (LightboxCaption, MatrixBox).

| Old | New |
|---|---|
| Components::IdentifyFilterForm | Views::Controllers::Observations::Identify::Form |
| Components::ImageEditForm | Views::Controllers::Observations::Images::Form |
| Components::ImportedSourceBanner | Views::Controllers::Observations::ImportedSourceBanner |
| Components::NamingForm | Views::Controllers::Observations::Namings::Form |
| Components::ObservationForm | Views::Controllers::Observations::Form |
| Components::ObserverQuestionForm | Views::Controllers::Observations::Emails::Form |

Notes-panel mixin follows ObservationForm:
`app/components/observation_form/notes_panel.rb` →
`app/views/controllers/observations/form/notes_panel.rb`
(class becomes `Views::Controllers::Observations::Form::NotesPanel`).

Tests moved and namespaced to match. 364 tests pass (34 view +
330 controller, including all `observations/*` nested controllers).

The `IdentifyFilterForm` keeps its explicit `id: "identify_filter"`
override — the navbar search Stimulus controllers depend on that
specific DOM id, and the new all-segments heuristic would yield
`observation_identify_form` otherwise. `NamingForm` keeps its
dynamic per-naming id (`obs_<id>_naming_<n>_form`) so multiple
namings on one obs can render without colliding.

Stays in `app/components/` (audit was imprecise — rendered by
reusable components):
- LightboxObservationTitle (used by LightboxCaption)
- MatrixBoxTitle (used by MatrixBox)
- Carousel components (carousel_thumbnail, form_carousel,
  form_carousel_item) — planned reuse in other model forms.
