+++ app/views/controllers/observations/namings/new.html.erb
@@ -12,7 +12,7 @@ add_context_nav(naming_form_new_tabs(obs: @observation))
     </div>
 
     <div class="mt-3">
-      <%= render(Components::NamingForm.new(
+      <%= render(Views::Controllers::Observations::Namings::Form.new(
             @naming,
             observation: @observation,
             vote: @vote,
